add ui-router-ng2 (#91)
* Apply Styleguide Rules 04-* Application Structure (#858)

* Apply Styleguide Rules 04-* Application Structure

Applies the Styleguide rules for the application structure.
Fixes #857.

* Update barrel imports

Updates the barrel imports to the `/index` syntax. Furthermore,
corrects the template and style urls of the components.

* Update README

Updates the directory structure in the README to reflect the
recent changes to the application folder structure to conform
to the styleguide reccomendations.

* Document tools section (#865)

* Document tools section

Documents the tools section of the seed. It provides a general
overview of the provided configuration, manual typings, tasks and
utilities.

* Update tools documentation

Updates the tools documentation by adding descriptions for the
provides tasks and utilities

* Document project and seed configuration (#866)

* Document project and seed configuration

Adds inline documentation for the project.config.ts and seed.config.ts using
jsdoc notation (http://usejsdoc.org/). Fixes #861.

* Fixes indentation and misspellings

Fixes the indentation for the constants declaration and fixes remaining
misspellingsin the documentation.

* Document sources in src/client (#878)

Provides inline documentation for the TypeScript files
within `src/client` using the jsdoc notation.
Contributes towards fixing #867.

* Document sources in tools/utils (#879)

Provides inline documentation for the TypeScript files
within `tools/utils` using the jsdoc notation.
Contributes towards fixing #867.

* added --b flag to prevent browser from open on serve.dev (#880)

* docs(readme): fix typo

* Document sources in tools/tasks (#884)

Provides inline documentation for the TypeScript files
within `tools/tasks` using the jsdoc notation.
Contributes towards fixing #867.

* Add gulp-replace package and use it to strip the global 'use strict' added by reflect-metadata (#802). (#885)

* chore(docs): drop tsdoc (#881)

Fix #830

* chore(docs): drop tsdoc (#881)

Fix #830

* feat(http): add sample http service calls

Fix #839

* Configurable client folder (#889)

With that it's possible to keep different clients in the project

* docs(contributors): update the list of contributors

* Apply `moduleId` to @Component annotation (#891)

* Apply `moduleId` to @Component annotation

Sets the `moduleId` property of the @Component annotation to enable
`templateUrl` and `styleUrls` to be only the file name without the
surrounding path.

* Update gulp-inline-ng2-template configuration

Updates the gulp-inline-ng2-template configuration to
use relative paths.

* Update gulp-inline-ng2-template configuration

Updates the gulp-inline-ng2-template configuration
in the build.js.prod task to use relative paths.

* Indent inconsistent (#899)

* Indent inconsistent

* Update about.component.e2e-spec.ts

* unnecessary selector

* Update about.component.e2e-spec.ts

* Update home.component.e2e-spec.ts

* Electron livereload (#83)

* electron-livereload: init

* removed vim files

* electron-livereload: remove not needed comment

* electron-livereload: remove prereq comment in README.md && conditionally inject electron-connect in index.html

* electron-livereload: changed double to single quotes

* chore: cleanup

* chore: update contributors table

* refactor: contain testing setups with their respective frameworks

* chore: improve merge upstream workflow

closes https://github.com/NathanWalker/angular2-seed-advanced/issues/84

* chore: testing

* readme: improved project usage instructions

* ns-livesync-device: added scripts start.livesync.android.device && start.livesync.ios.device (#87)

* add ui-router-ng2