fix: nativescript build template interpolation
This disables the ES template literal interpolation `${ ... }` that was
causing the tns js build to fail due to comments containing that pattern
and the corresponding interpolation param not being found.

Specifically this bit of commented out code in the ns-app.service file:
// this.log.info(`NSAppComponent translate.onLangChange(${args.lang})`);

I disabled it in all calls to the gulp template plugin for consistancy.

Related lodash issue: https://github.com/lodash/lodash/issues/399